made authorized only navbar and updated imports
