+++ web-client/src/pages/dashboard.js
@@ -1,32 +1,15 @@
-import React, { useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import NavBar from '../navigationBar';
+
+import NavBar from '../Components/NavBars/authorizedNavBar';
 import {Button} from 'react-bootstrap';
-import { useUserAuth } from '../firebase/UserAuthContext';
 
 export default function Dashboardashboard() {
-  const [error, setError] = useState("");
-  const { logOut } = useUserAuth();
-  const navigate = useNavigate();
 
-  const handleLogOut = async (e) => {
-    e.preventDefault();
-    setError("");
-    try {
-      await logOut();
-      navigate("/");
-    } catch (err) {
-      setError(err.message);
-    }
-  };
 
   return (
     <>
     <NavBar/>
     <div>Dashboard</div>
-    <Button variant="primary" onClick={handleLogOut}> 
-     Log Out 
-    </Button>
+
     </>
   )
 }
